+++ src/Users/Policies/UserPolicy.php
@@ -119,6 +119,10 @@ class UserPolicy
      */
     public function update(User $auth, User $user)
     {
+        if ($auth->role->is(Roles::Administrator) && $user->isMyself) {
+            return true;
+        }
+
         if (!is_null($user->deleted_at) || $user->isProtected) {
             return false;
         }