+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Update:
 - Added Enraiged\Geo system for handling countries,regions,timeszones
 - Corrections,improvements to the User Roles enum
 - Corrections,improvements to the Profile Salutations enum
+- Improvements to the user context attributes
 - Refactored Enraiged\NetworkAddresses namespace to Enraiged\Network
 
 ---