+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Update:
 - Corrections,improvements to the User Roles enum
 - Corrections,improvements to the Profile Salutations enum
 - Improvements to the user context attributes
+- Rebuild the user form template; Readded original sections
 - Refactored Enraiged\NetworkAddresses namespace to Enraiged\Network
 
 ---